feat: :sparkles: sideNav 관련 파일 분리 및 이동
